Fall back to local copy for test4theory.js Closes #2
index.html tries first to load the CDN version and if it fails, it loads the
local copy. This will allow to update clients JS without having to ask them to
do nothing special, while at the same time there is an off-line solution if
the CDN is not working.